Make avatar resonsive
