+++ src/frontend/src/lib/components/address-book/AddContactStep.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { notEmptyString } from '@dfinity/utils';
-	import Avatar from '$lib/components/address-book/Avatar.svelte';
+	import Avatar from '$lib/components/contact/Avatar.svelte';
 	import ContactForm from '$lib/components/address-book/ContactForm.svelte';
 	import Button from '$lib/components/ui/Button.svelte';
 	import ButtonCancel from '$lib/components/ui/ButtonCancel.svelte';
